+++ app/ai_tools/delete_day.rb
@@ -5,7 +5,8 @@ module AITools
 
     def execute(day_id:)
       with_rescues do
-        day = Day.find_by(id: day_id)
+        next require_tour! if @tour.nil?
+        day = @tour.days.find_by(id: day_id)
         next bail("Day not found", code: "day_not_found") unless day
         day.destroy!
         ok(deleted_day_id: day_id)
